PL-7914 Endringer i brev-api for inntektsendringsbrev for å ta høyde for endringer i ny api-versjon
