+++ pensjon-brevbaker-api-model/gradle.properties
@@ -1,1 +1,1 @@
-version=141
+version=142
